Use data-uris for image assets in emails
Fixes GH-4038
